Created a better logger way (ScopedLogger) now logging as : [Namespace.ClassName.Method] MESSAGE. Fixed an error in the license (Was MIT, but the project is under Apache2.0 license!). Make some optimizations for type convert (Vector2, Vector3, Matrix4x4). Worked on the camera system. Worked on the map chunk system. Worked on the layer system. Added CameraComponent. Added CameraService. Unified the Dialog and Menu inside specific services (Can be found in UiServices). Started to work on rendering during RTP time. And other thing that I don't remember.

+++ RPGCreator.Core/Managers/RTP/BrushManagers/Brushs/IBrush.cs
@@ -3,14 +3,14 @@
 // RPG Creator - Open-source RPG Engine.
 // (c) 2025 Ward
 // 
-// This file is part of RPG Creator and is distributed under the MIT License.
-// You are free to use, modify, and distribute this file under the terms of the MIT License.
+// This file is part of RPG Creator and is distributed under the Apache 2.0 License.
+// You are free to use, modify, and distribute this file under the terms of the Apache 2.0 License.
 // See LICENSE for details.
 // 
 // ---
 // 
-// Ce fichier fait partie de RPG Creator et est distribué sous licence MIT.
-// Vous êtes libre de l'utiliser, de le modifier et de le distribuer sous les termes de la licence MIT.
+// Ce fichier fait partie de RPG Creator et est distribué sous licence Apache 2.0.
+// Vous êtes libre de l'utiliser, de le modifier et de le distribuer sous les termes de la licence Apache 2.0.
 // Voir LICENSE pour plus de détails.
 // 
 // Contact:
@@ -42,7 +42,7 @@ namespace RPGCreator.Core.Managers.RTP.BrushManagers.Brushs
                 return false;
             }
 
-            int cellSize = mapInstance.Definition.GridParameter.CellWidth;
+            float cellSize = mapInstance.Definition.GridParameter.CellWidth;
             float horizontalCells = mapInstance.Definition.Size.Width;
             float verticalCells = mapInstance.Definition.Size.Height;
